third_party/blink/web_tests/platform/linux/virtual/sanitizer-api/external/wpt/sanitizer-api/sanitizer-boolean-defaults.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] data attributes
  assert_equals: expected (boolean) true but got (undefined) undefined
Harness: the test ran to completion.
